Secure Password Generator v1.4
# SecurePasswordGenerator
It's a open source Secure Password Generator to help with very secure passwords.

v1.4 Change log : 
Now it saves the passwords on a folder, Now it asks you how many letters you want the password to have, for what to be and how many and soon a cool print! ( Thank you very much Raluvy95 for the help )

v1.3 Change log :
Now it opens a CMD with the password,
 The file is smaller,
 It now generates from default a 32 digit password ( you can change it if you need )

v1.2 Change log :
Updated the generated passwords from 15 characters to 16 characters
, updated the print

v1.1 Change log :
Added a welcome message.

v1.0 Change log:

Created the Generator and made it working.